app/main.py
from fastapi import FastAPI, WebSocket, Depends, HTTPException, Security
from fastapi.responses import HTMLResponse
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
import logging
from logging.handlers import RotatingFileHandler
import asyncio
from typing import List, Dict

app = FastAPI()

@app.post("/upload-pdf/")
async def upload_pdf():
    return {"message": "Processed"}

@app.post("/save-template/")
async def save_template():
    return {"message": "Template saved successfully."}

@app.get("/get-template/")
async def get_template(template_name: str):
    # Placeholder for actual template retrieval logic
    return {"template_name": template_name, "x": 1, "y": 2, "z": 3}

@app.post("/process-ocr/")
async def process_ocr(template_name: str, rectangles: List[Dict[str, int]]):
    # Simulate processing based on 'template_name' and 'rectangles'
    # This is a placeholder. Replace with actual processing logic as needed.
    return {"message": "OCR processed successfully!"}

# Ensure these endpoints exist
@app.post("/token")
async def authenticate_user(form_data: OAuth2PasswordRequestForm = Depends()):
    # Implement actual authentication logic here
    return {"access_token": "some_token", "token_type": "bearer"}

@app.get("/secure-endpoint")
async def secure_endpoint(token: str = Security(OAuth2PasswordBearer(tokenUrl="/token"))):
    # Implementation goes here, token should be checked for validity
    return {"message": "Secure content"}

# Setup advanced logging configuration
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', handlers=[RotatingFileHandler("app_logs.log", maxBytes=1000000, backupCount=3)])
logger = logging.getLogger(__name__)

html = """
<!DOCTYPE html>
<html>
    <head>
        <title>OCR Processing Status</title>
    </head>
    <body>
        <h1>WebSocket Test</h1>
        <button onclick='connectWebSocket()'>Connect</button>
        <script>
            function connectWebSocket() {
                var ws = new WebSocket('ws://localhost:8000/ws');
                ws.onmessage = function(event) {
                    var message = event.data;
                    document.body.append(message + '<br>');
                };
                ws.onerror = function(event) {
                    console.error('WebSocket error observed:', event);
                };
                ws.onclose = function(event) {
                    console.log('WebSocket connection closed', event);
                };
            }
        </script>
    </body>
</html>
"""

@app.get("/")
async def get():
    return HTMLResponse(html)

@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    try:
        await websocket.accept()
        for i in range(100):
            await websocket.send_text(f'Message {i}')
            await asyncio.sleep(1)
    except Exception as e:
        logger.error(f"WebSocket error: {e}")
    finally:
        await websocket.close()
